pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cybersec-agents"
version = "1.0.0"
description = "Specialized AI agents for cybersecurity tasks using the Camel AI framework"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = ["ai", "cybersecurity", "agents", "camel", "llm"]
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "camel-ai>=0.2.29",
    "openai>=1.0.0",
    "agentops>=0.1.0",
    "pandas>=1.3.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.12.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=0.19.0",
    "requests>=2.32.0",
    "numpy>=1.26.0",
    "flask>=2.3.0",
    "discord.py>=2.3.0",
    "scikit-learn>=1.3.0",
    "nltk>=3.8.1",
    "anthropic>=0.3.0",
    "dependency-injector>=4.41.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.3.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "autoflake>=2.0.0",
]
docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.0.0",
    "myst-parser>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/cybersec-agents"
Documentation = "https://yourusername.github.io/cybersec-agents"
Repository = "https://github.com/yourusername/cybersec-agents.git"
Issues = "https://github.com/yourusername/cybersec-agents/issues"

[tool.black]
line-length = 88
target-version = ["py310", "py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = false
disallow_untyped_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
ignore_missing_imports = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = [
    "camel-ai.*",
    "agentops.*",
    "pandas.*",
    "matplotlib.*",
    "seaborn.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=cybersec_agents --cov-report=term-missing"
testpaths = [
    "tests",
]

[tool.coverage.run]
source = ["cybersec_agents"]
omit = [
    "tests/*",
    "docs/*",
    "examples/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
]
